tc358743: Add support for platforms without IRQ line
The existing driver relied on having the interrupt line
connected from the TC358743 back to the SoC.
Add the option to poll the interrupt status register via
a timer if interrupts is not defined.

interrupts is listed as an optional property in DT binding,
but in reality the driver didn't work without it.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>