staging: bcm2835-camera: Remove V4L2/MMAL buffer remapping
The MMAL and V4L2 buffers had been disassociated, and
linked on demand. Seeing as both are finite and low in
number, and we now have the same number of each, link them
for the duration.
This removes the complexity of maintaining lists as the
struct mmal_buffer context comes back from the VPU, so we
can directly link back to the relevant V4L2 buffer.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>